cleanup, consolidation of tests
